arbeit

+++ app/src/main/java/com/example/asynctascretrofit/model/ForecastDays/ForCastItemNumbeZeroFourth.kt
@@ -3,7 +3,7 @@ package com.example.asynctascretrofit.model.ForecastDays
 data class ForCastItemNumbeZeroFourth(
     val dt : Int,
     val visibility : Int,
-    val pop : Int,
+    val pop : Double,
     val dt_txt : String,
     val  main : ForCastMainFifth
